Mostly automated py2 -> py3 breaking scripts changes.
Run 2to3 fixers:
- unicode
- dict
- xrange
- long

Also manually fixed the abc metaclass usage to instead use `abc.ABC`.

+++ Scripts/Python/nglnUrwinBrain.py
@@ -325,7 +325,7 @@ class nglnUrwinBrain(ptResponder):
         PtDebugPrint("nglnUrwinBrain: Generated a valid spawn time: %d" % (firstTime))
         spawnTimes = [firstTime]
 
-        while isinstance(spawnTimes[-1], long):
+        while isinstance(spawnTimes[-1], int):
             randnum = random.randint(kMinimumTimeBetweenSpawns, kMaximumTimeBetweenSpawns)
             newTime = spawnTimes[-1] + randnum
             if newTime < endOfToday:
